+++ lib/features/group_profile/presentation/widgets/group_profile_drawer.dart
@@ -1,0 +1,93 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_pecha/core/widgets/error_state_widget.dart';
+import 'package:flutter_pecha/features/group_profile/presentation/providers/group_profile_providers.dart';
+import 'package:flutter_pecha/features/group_profile/presentation/widgets/group_profile_body.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+class GroupProfileDrawer extends ConsumerWidget {
+  final String groupId;
+
+  const GroupProfileDrawer({super.key, required this.groupId});
+
+  static Future<void> show(BuildContext context, String groupId) {
+    return showModalBottomSheet(
+      context: context,
+      isScrollControlled: true,
+      backgroundColor: Colors.transparent,
+      isDismissible: true,
+      enableDrag: true,
+      barrierColor: Colors.black.withValues(alpha: 0.5),
+      builder: (_) => GroupProfileDrawer(groupId: groupId),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final profileAsync = ref.watch(groupProfileProvider(groupId));
+    final isDark = Theme.of(context).brightness == Brightness.dark;
+
+    return DraggableScrollableSheet(
+      initialChildSize: 0.75,
+      minChildSize: 0.4,
+      maxChildSize: 0.95,
+      builder: (context, scrollController) {
+        return Container(
+          decoration: BoxDecoration(
+            color: Theme.of(context).scaffoldBackgroundColor,
+            borderRadius: const BorderRadius.vertical(
+              top: Radius.circular(20),
+            ),
+          ),
+          child: Column(
+            children: [
+              _buildDragHandle(context),
+              Expanded(
+                child: profileAsync.when(
+                  data: (either) => either.fold(
+                    (failure) => Center(
+                      child: ErrorStateWidget(
+                        error: failure,
+                        onRetry: () =>
+                            ref.invalidate(groupProfileProvider(groupId)),
+                      ),
+                    ),
+                    (profile) => GroupProfileBody(
+                      profile: profile,
+                      isDark: isDark,
+                      scrollController: scrollController,
+                      onSeriesTap: () => Navigator.of(context).pop(),
+                    ),
+                  ),
+                  loading: () =>
+                      const Center(child: CircularProgressIndicator()),
+                  error: (error, _) => Center(
+                    child: ErrorStateWidget(
+                      error: error,
+                      onRetry: () =>
+                          ref.invalidate(groupProfileProvider(groupId)),
+                    ),
+                  ),
+                ),
+              ),
+            ],
+          ),
+        );
+      },
+    );
+  }
+
+  Widget _buildDragHandle(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(top: 12, bottom: 8),
+      child: Container(
+        width: 40,
+        height: 4,
+        decoration: BoxDecoration(
+          color:
+              Theme.of(context).colorScheme.onSurface.withValues(alpha: 0.3),
+          borderRadius: BorderRadius.circular(2),
+        ),
+      ),
+    );
+  }
+}
